ARCHITECTURAL/ArchiGPT_0.2/datasets/initial/10_RentYourExpert/rentyourexpert-main/backend/catalogue_ms/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
mysql-connector-python
Werkzeug==2.2.2